feat: 完成原料购买界面和基础HUD系统
- 新增原料商店槽位管理系统 (IngredientStoreSlot/IngredientStoreSlotManager)
  - 实现原料购买界面的完整逻辑
  - 支持序列化槽位配置和动态激活
  - 添加数量增减按钮和输入框交互
  - 集成购物车管理和价格计算

- 增强HUD系统
  - 添加天数和声望显示功能
  - 实时更新游戏状态信息
  - 优化UI面板管理逻辑

- 改进交互系统
  - 添加观察冷却时间机制
  - 优化相机视角切换流程
  - 完善交互状态管理

- 重构资源组织结构
  - 重新整理场景文件结构
  - 优化脚本管理器分类
  - 更新资源配置和交互数据

- 其他优化
  - 将装备和原料数据脚本中的价格改成int

+++ 2025gameoff/Assets/Scripts/Data/EquipmentDataSO.cs
@@ -9,7 +9,7 @@ public class EquipmentDataSO : ScriptableObject
     [Tooltip("装备名称")]
     public string equipmentName;
     [Tooltip("售价")]
-    public float equipmentPrice;
+    public int equipmentPrice;
     [TextArea]
     public string description;
 }
